Changed default bogus values to have only one value
